% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebirdst-deprecated.R
\name{abundance_palette-deprecated}
\alias{abundance_palette-deprecated}
\title{eBird Status and Trends color palettes for mapping}
\usage{
abundance_palette(n,
                        season = c("weekly", "breeding",
                                   "nonbreeding",
                                   "migration",
                                   "prebreeding_migration",
                                   "postbreeding_migration",
                                   "year_round"))
}
\arguments{
\item{n}{integer; the number of colors to be in the palette.}

\item{season}{character; the season to generate colors for or "weekly" to
get the color palette used in the weekly abundance animations.}
}
\value{
A character vector of hex color codes.
}
\description{
This deprecated function has been replaced by \code{\link{ebirdst_palettes}}.
Both functions generate color palettes used for the eBird Status and Trends
relative abundance maps.
}
\seealso{
\code{\link{ebirdst_palettes}} \code{\link{ebirdst-deprecated}}
}
\keyword{internal}
