% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stemhelper-ppms.R
\name{balanced.binary.sample}
\alias{balanced.binary.sample}
\title{Internal balanced binary sample function}
\usage{
balanced.binary.sample(binary.outcome, x, y, xlim = NA, ylim = NA,
  nx = 50, ny = 50, min.class = 0.25, neg.size = NA)
}
\description{
Internal balanced binary sample function
}
