% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebirdst-ppms.R
\name{bernoulli_dev}
\alias{bernoulli_dev}
\title{Bernoulli deviance}
\usage{
bernoulli_dev(obs, pred)
}
\arguments{
\item{obs}{numeric; observed values.}

\item{pred}{numeric; predicted values.}
}
\value{
A named numeric vector with three elements: model deviance, mean
deviance, and deviance explained.
}
\description{
Bernoulli deviance
}
\examples{
obs <- c(1, 1, 1, 0, 0, 0)
pred <- c(0.9, 0.8, 0.7, 0.3, 0.1, 0.2)
ebirdst:::bernoulli_dev(obs, pred)
}
