% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stemhelper-mapping.R
\name{calc_effective_extent}
\alias{calc_effective_extent}
\title{Map extent of estimation calculated from subset of centroids}
\usage{
calc_effective_extent(st_extent, pis = NA, pds = NA)
}
\description{
Map extent of estimation calculated from subset of centroids
}
