% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stemhelper-mapping.R
\name{combine_layers}
\alias{combine_layers}
\title{Combine zero layers and abundance to create a single layer}
\usage{
combine_layers(path, week)
}
\description{
Combine zero layers and abundance to create a single layer
}
