% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stemhelper-plotting.R
\name{convert_classes}
\alias{convert_classes}
\title{Internal function for converting cover class names to readable}
\usage{
convert_classes(cov_names, by_cover_class = FALSE, pretty = FALSE)
}
\description{
Internal function for converting cover class names to readable
}
