% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebirdst-loading.R
\name{data_st_subset}
\alias{data_st_subset}
\title{Spatiotemporal subsetter for STEM result data objects}
\usage{
data_st_subset(data, st_extent)
}
\arguments{
\item{data}{data.frame or SpatialPointsDataFrame; originating from results.}

\item{st_extent}{list; st_extent list object}
}
\value{
Subset of input data as same type.
}
\description{
Internal function that takes a data.frame or SpatialPointsDataFrame and
a st_extent list and returns a subset of the data object. Currently
designed to handle either a 'rectangle' as defined by a lat/lon bounding
box or a 'polygon' as defined by a SpatialPolygon* object. The function
will use temporal information if provided. The t.min and t.max objects in
the \code{st_extent} list are currently able to wrap time around the year
(e.g., t.min = 0.9 and t.max = 0.1 is acceptable).
}
\examples{
\dontrun{

# define species STEM results location and load pis
sp_path <- "path to species STEM results"
pis <- load_pis(sp_path)

# define st_extent list
ne_extent <- list(type = "rectangle",
                  lat.min = 40,
                  lat.max = 47,
                  lon.min = -80,
                  lon.max = -70,
                  t.min = 0.425,
                  t.max = 0.475)

#subset pis with st_extent list
data_st_subset(data = pis, st_extent = ne_extent, use_time = TRUE)
}
}
