% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{date_to_st_week}
\alias{date_to_st_week}
\title{Get the Status and Trends week that a date falls into}
\usage{
date_to_st_week(dates, version = 2022)
}
\arguments{
\item{dates}{a vector of dates.}

\item{version}{One of \code{2021} for the date scheme used for the 2021 and prior
data releases or \code{2022} for the date scheme used in the 2022 and subsequent
releases.  Default is \code{2022}.}
}
\value{
An integer vector of weeks numbers from 1-52.
}
\description{
Get the Status and Trends week that a date falls into
}
\examples{
d <- as.Date(c("2016-04-08", "2018-12-31", "2014-01-01", "2018-09-04"))
date_to_st_week(d)
}
