% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebirdst-loading.R
\name{download_data}
\alias{download_data}
\title{Downloads data package for a single species or for the example data}
\usage{
download_data(path, species, example_data = FALSE)
}
\arguments{
\item{path}{character; directory for data to be downloaded to}

\item{species}{character; either a valid six letter code or "example_data"}
}
\value{
Full path to results.
}
\description{
This function will download the entire data package for a single species to a
specified path.
}
\examples{
#' \dontrun{
path <- "~/tmp"
species <- "example_data"

download_data(path, species)
}
}
