% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_ebird_trends}
\alias{download_ebird_trends}
\title{Download eBird Trends Data Products}
\usage{
download_ebird_trends(
  path = ebirdst_data_dir(),
  force = FALSE,
  show_progress = TRUE
)
}
\arguments{
\item{path}{character; directory to download the data to. The downloaded file
containing the trends data will be placed in a sub-directory of this
directory named for the data version year, e.g. "2021" for the 2007-2021
Trends Data Products. Defaults to a persistent data directory, which can be
found by calling \code{ebirdst_data_dir()}.}

\item{force}{logical; if the data have already been downloaded, should a
fresh copy be downloaded anyway.}

\item{show_progress}{logical; whether to print download progress information.}
}
\value{
Path to the trends data file.
}
\description{
Download eBird Trends Data Products parquet file for all species. Accessing
Status and Trends data requires an access key, consult
\code{\link[=set_ebirdst_access_key]{set_ebirdst_access_key()}} for instructions on how to obtain and store this
key. Use \code{\link[=load_trends]{load_trends()}} to load the data into the R session after
downloading it or the parquet file can be loaded directly using
\link[arrow:read_parquet]{read_parquet()} from the \code{arrow} package.
}
\examples{
\dontrun{
path <- download_ebird_trends()
}
}
