% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebirdst-deprecated.R
\name{ebirdst-deprecated}
\alias{ebirdst-deprecated}
\alias{abundance_palette}
\title{Deprecated functions in package \pkg{ebirdst}.}
\usage{
abundance_palette(
  n,
  season = c("weekly", "breeding", "nonbreeding", "migration", "prebreeding_migration",
    "postbreeding_migration", "year_round")
)
}
\description{
The functions listed below are deprecated and support for them
will eventually be dropped.
Help pages for deprecated functions are
available at \code{help("<function>-deprecated")}.
}
\section{\code{abundance_palette}}{

For \code{abundance_palette}, use \code{\link{ebirdst_palettes}}
}

\keyword{internal}
