% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebirdst-package.R
\docType{package}
\name{ebirdst-package}
\alias{ebirdst}
\alias{ebirdst-package}
\title{\code{ebirdst}: Tools to Load, Map, Plot, and Analyze eBird Status and Trends
Data Products}
\description{
Tools to load, map, plot, and analyze \href{https://science.ebird.org/status-and-trends}{eBird Status and Trends} data products
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ebird.github.io/ebirdst/}
  \item \url{https://github.com/ebird/ebirdst}
  \item Report bugs at \url{https://github.com/ebird/ebirdst/issues}
}

}
\author{
\strong{Maintainer}: Matthew Strimas-Mackey \email{mes335@cornell.edu} (\href{https://orcid.org/0000-0001-8929-7776}{ORCID})

Authors:
\itemize{
  \item Matthew Strimas-Mackey \email{mes335@cornell.edu} (\href{https://orcid.org/0000-0001-8929-7776}{ORCID})
  \item Shawn Ligocki \email{sligocki@cornell.edu}
  \item Myles Stokowski \email{mas892@cornell.edu}
  \item Tom Auer \email{mta45@cornell.edu} (\href{https://orcid.org/0000-0001-8619-7147}{ORCID})
  \item Daniel Fink \email{df36@cornell.edu} (\href{https://orcid.org/0000-0002-8368-1248}{ORCID})
}

Other contributors:
\itemize{
  \item Cornell Lab of Ornithology [copyright holder]
}

}
\keyword{internal}
