% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebirdst-package.R
\docType{package}
\name{ebirdst}
\alias{ebirdst}
\title{\code{ebirdst}: Tools to Load, Map, Plot, and Analyze eBird Status and Trends
Data Products}
\description{
Tools to load, map, plot, and analyze \href{https://ebird.org/science/status-and-trends}{eBird Status and Trends} data products
}
