% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebirdst-loading.R
\name{ebirdst_copy}
\alias{ebirdst_copy}
\title{Copy eBird Status and Trends data from the mount to your local drive}
\usage{
ebirdst_copy(
  species,
  tifs_only = TRUE,
  path = rappdirs::user_data_dir("ebirdst")
)
}
\arguments{
\item{species}{character; a single species given as a scientific name, common
name or six-letter species code (e.g. woothr). The full list of valid
species is can be viewed in the \link{ebirdst_runs} data frame included in this
package. To download the example dataset, use "example_data".}

\item{tifs_only}{logical; whether to only download the GeoTIFFs for
abundance and occurrence (the default), or download the entire data
package, including data for predictor importance, partial dependence, and
predictive performance metrics.}

\item{path}{character; directory to download the data to. All downloaded
files will be placed in a sub-directory of this directory named according
to the unique run ID associated with this species. Defaults to a persistent
data directory, which can be found by calling
rappdirs::user_data_dir("ebirdst")).}
}
\description{
Copy eBird Status and Trends data from the mount to your local drive
}
