% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{ebirdst_data_dir}
\alias{ebirdst_data_dir}
\title{Path to eBird Status and Trends data download directory}
\usage{
ebirdst_data_dir()
}
\value{
The path to the data download directory.
}
\description{
Identify and return the path to the default download directory for eBird
Status and Trends data products. This directory can be defined by setting the
environment variable \code{EBIRDST_DATA_DIR}, otherwise the directory returned by
\code{tools::R_user_dir("ebirdst", which = "data")} will be used.
}
\examples{
ebirdst_data_dir()
}
