% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebirdst-palettes.R
\name{ebirdst_palettes}
\alias{ebirdst_palettes}
\title{eBird Status and Trends color palettes for mapping}
\usage{
ebirdst_palettes(
  n,
  type = c("weekly", "breeding", "nonbreeding", "migration", "prebreeding_migration",
    "postbreeding_migration", "year_round", "trends")
)
}
\arguments{
\item{n}{integer; the number of colors to be in the palette.}

\item{type}{character; the type of color palette: "weekly" for the weekly
relative abundance, "trends" for trends color palette, and a season name for
the seasonal relative abundance. Note that for trends a diverging palette
is returned, while all other palettes are sequential.}
}
\value{
A character vector of hex color codes.
}
\description{
Generate the color palettes used for the eBird Status and Trends relative
abundance and trends maps.
}
\examples{
# breeding season color palette
ebirdst_palettes(10, type = "breeding")
}
