% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ebirdst_predictor_datasets}
\alias{ebirdst_predictor_datasets}
\title{eBird Status and Trends predictors datasets}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 37 rows and 4 columns.
}
\usage{
ebirdst_predictor_datasets
}
\description{
eBird Status and Trends predictors datasets
}
\keyword{datasets}
