% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ebirdst_predictor_descriptions}
\alias{ebirdst_predictor_descriptions}
\title{eBird Status and Trends predictors descriptions}
\format{
A data frame with 37 rows and 4 columns
\itemize{
\item \code{dataset}: dataset name.
\item \code{predictor}: predictor name or, if multiple variables are derived from
this dataset, the pattern used to generate the names.
\item \code{description}: detailed description of the dataset or variable.
\item \code{reference}: a reference to consult for further information on the dataset.
}
}
\usage{
ebirdst_predictor_descriptions
}
\description{
Details on the eBird Status and Trends predictor variables or, for variables
all derived from the same dataset, details on the dataset.
}
\keyword{datasets}
