% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ebirdst_predictors}
\alias{ebirdst_predictors}
\title{eBird Status and Trends predictors}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 150 rows and 4 columns.
}
\usage{
ebirdst_predictors
}
\description{
eBird Status and Trends predictors
}
\keyword{datasets}
