% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebirdst-subset.R
\name{ebirdst_subset}
\alias{ebirdst_subset}
\alias{ebirdst_subset.data.frame}
\alias{ebirdst_subset.sf}
\alias{ebirdst_subset.Raster}
\title{Subset eBird Status and Trends data spatiotemporally}
\usage{
ebirdst_subset(x, ext)

\method{ebirdst_subset}{data.frame}(x, ext)

\method{ebirdst_subset}{sf}(x, ext)

\method{ebirdst_subset}{Raster}(x, ext)
}
\arguments{
\item{x}{eBird Status and Trends data to subset; either a Raster object with
52 layers (one for each week) or a \code{data.frame} with PI or PD data.}

\item{ext}{\link{ebirdst_extent} object; the spatiotemporal extent to
filter the data to.}
}
\value{
eBird Status and Trends data in the same format as the input data.
}
\description{
Spatiotemporally subset the raster or tabular eBird Status and Trends data.
The spatiotemporal extent should be defined using \code{\link[=ebirdst_extent]{ebirdst_extent()}}.
}
\section{Methods (by class)}{
\itemize{
\item \code{data.frame}: PI or PD data

\item \code{sf}: PI or PD data as an \code{sf} object

\item \code{Raster}: Status and Trends rasters
}}

\examples{
\dontrun{

# bbox for southern michigan in may
bb_vec <- c(xmin = -86.6, xmax = -82.2, ymin = 41.5, ymax = 43.5)
e <- ebirdst_extent(bb_vec, t = c("05-01", "05-31"))

# download and load example data
sp_path <- download_data("example_data")
pis <- load_pis(sp_path)
abd <- load_raster(product = "abundance_umean", sp_path)

# subset
abd_ss <- ebirdst_subset(abd, ext = e)
pis_ss <- ebirdst_subset(pis, ext = e)

}
}
