% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ebirdst_trends_runs}
\alias{ebirdst_trends_runs}
\title{Data frame of species with eBird Trends Data Products}
\format{
A data frame with 11 variables:
\describe{
\item{run_name}{Combination of species code and run name.}
\item{species_code}{The alphanumeric eBird species code uniquely
identifying the species.}
\item{common_name}{The English common name of the species.}
\item{season}{Season that the trend was estimated for: breeding,
nonbreeding, or resident.}
\item{region}{The geographic region that the trend model was run for. Note
that broadly distributed species (e.g. Barn Swallow) will only have
trend estimates for a regional subset of their full range.}
\item{start_date}{The start date (\code{MM-DD} format) of the season for which
the trend was estimated.}
\item{end_date}{The end date (\code{MM-DD} format) of the season for which the
trend was estimated.}
\item{start_year}{The start date of the trend time period.}
\item{end_year}{The end date of the trend time period.}
\item{rsquared}{R-squared value comparing the actual and estimated trends
from the simulations.}
\item{beta0}{The slope of a linear model fitting actual vs. estimated
trends (\code{actual ~ estimated}) for the simulations. Positive values
of \code{beta0} indicate that the models are systematically
\emph{underestimating} the simulated trend for this species.}
}
}
\usage{
ebirdst_trends_runs
}
\description{
This data frame contains a list of species for which eBird Trends Data
Products are available including model-level information for the trend model.
The two predictive performance metrics (\code{rsquared} and \code{beta0}) are based
on a comparison of actual and estimated percent per year trends for a suite
of simulations (see Fink et al. 2023 for further details).
}
\keyword{datasets}
