% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{ebirdst_version}
\alias{ebirdst_version}
\title{eBird Status and Trends Data Products version}
\usage{
ebirdst_version()
}
\value{
A list with two components:
for the eBird Status Data Products
\itemize{
\item \code{status_version_year}: version year for the eBird Status Data Products
\item \code{trends_version_year} version year for the eBird Trends Data Products
}
}
\description{
Identify the version of the eBird Status and Trends Data Products that this
version of the R package works with. Versions are defined by the year that
all model estimates are made for.
}
\examples{
ebirdst_version()
}
