% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stemhelper-loading.R
\name{get_sinu_ext}
\alias{get_sinu_ext}
\title{Internal function for transforming st_extent to sinusoidal raster extent}
\usage{
get_sinu_ext(extent)
}
\description{
Internal function for transforming st_extent to sinusoidal raster extent
}
