% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{get_trends_path}
\alias{get_trends_path}
\title{Get the path to the trends data parquet file}
\usage{
get_trends_path(path = ebirdst_data_dir())
}
\arguments{
\item{path}{character; directory to download the data to. The downloaded file
containing the trends data will be placed in a sub-directory of this
directory named for the data version year, e.g. "2021" for the 2007-2021
Trends Data Products. Defaults to a persistent data directory, which can be
found by calling \code{ebirdst_data_dir()}.}
}
\value{
The path to the data package directory.
}
\description{
This helper function can be used to get the path to trends data parquet file.
This path can be passed to \code{\link[=load_trends]{load_trends()}} to load the data into the R
session or the parquet file can be loaded directly using
\link[arrow:read_parquet]{read_parquet()} from the \code{arrow} package.
}
\examples{
\dontrun{
# download the trends data
download_ebird_trends()

# get the path to the parquet file
path <- get_trends_path()
}
}
