% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stemhelper-ppms.R
\name{load_ppm_data}
\alias{load_ppm_data}
\title{Load Predictive Performance Metrics Data}
\usage{
load_ppm_data(path)
}
\arguments{
\item{path}{character; Full path to single species STEM results.}
}
\value{
list with data.frame of ppm_data and data.frame of interpolated
daily ensemble support values for both North America and South America.
}
\description{
Internal function that loads test data and filters it against ensemble
support data.
}
\examples{
\dontrun{

sp_path <- "path to species STEM results"

ppm_data <- load_ppm_data(sp_path)
}
}
\keyword{internal}
