% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stemhelper-loading.R
\name{load_summary}
\alias{load_summary}
\title{Summary file loader}
\usage{
load_summary(path)
}
\description{
Internal function used by load_pis() and load_pds() to get the stixel
summary information
}
