% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stemhelper-plotting.R
\name{loess_fit_and_predict}
\alias{loess_fit_and_predict}
\title{Loess fit and predict for cake_plot() PIs and PDs}
\usage{
loess_fit_and_predict(x, ext, input_data, type)
}
\arguments{
\item{x}{character; Predictor name.}

\item{ext}{list; st_extent list}

\item{input_data}{data.frame; Munged PIs or PD slopes.}

\item{type}{character; Either "PI" or "PD"}
}
\value{
Either smoothed PIs or smoother PD slopes.
}
\description{
Internal function that performs a loess fit and predict on both PIs and PDs
for smoothing each predictor.
}
\examples{
\dontrun{
# from cake_plot()
temporal smoothing of predictor trajectories
loesses <- lapply(X = unique(pd_slopes$predictor),
                  FUN = loess_fit_and_predict,
                  ext = st_extent,
                  input_data = pd_slopes,
                  type = "PD")
}
}
\keyword{internal}
