% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stemhelper-ppms.R
\name{lookup.grid.cell}
\alias{lookup.grid.cell}
\title{Internal lookup.grid.cell function}
\usage{
lookup.grid.cell(xxx, yyy, xlim = c(NA, NA), ylim = c(NA, NA), nx = 64,
  ny = 64, jitter = F)
}
\description{
Internal lookup.grid.cell function
}
