% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stemhelper-mapping.R
\name{maps_centroids}
\alias{maps_centroids}
\title{Map PI and PD centroid locations for species}
\usage{
maps_centroids(pis, pds, st_extent = NA, plot_pis = TRUE, plot_pds = TRUE,
  ...)
}
\description{
Map PI and PD centroid locations for species
}
