% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebirdst-loading.R
\name{parse_raster_dates}
\alias{parse_raster_dates}
\title{Parses the names attached to a Raster* from label_raster_stack()}
\usage{
parse_raster_dates(raster)
}
\arguments{
\item{raster}{RasterLayer, RasterStack, or RasterBrick; full or subser of
original eBird Status and Trends product raster GeoTiff.}
}
\value{
Date vector.
}
\description{
The label_raster_stack() function labels the dates of the estimate rasters
in the format of "XYYYY.MM.DD", because of constraints in the \code{raster}
packge. This function converts that character vector into an ISO compliant
Date vector.
}
\examples{
\dontrun{
rs <- stack("/path to GeoTiff/")
rs_labeled <- label_raster_stack(raster_data)
rs_dates <- parse_raster_dates(rs_labeled)
}
}
