% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stemhelper-ppms.R
\name{plot_binary_by_time}
\alias{plot_binary_by_time}
\title{Plot binary PPMs by time (week, month, custom)}
\usage{
plot_binary_by_time(path, metric = c("Kappa", "AUC", "Sensitivity",
  "Specificity"), st_extent = NA)
}
\description{
Plot binary PPMs by time (week, month, custom)
}
