% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebirdst-plotting.R
\name{plot_pis}
\alias{plot_pis}
\title{Plot predictor importances as barplots}
\usage{
plot_pis(path, pis, st_extent, by_cover_class = FALSE,
  num_top_preds = 50, return_top = FALSE, pretty_names = TRUE,
  print_plot = TRUE)
}
\arguments{
\item{path}{character; full path to single species eBird Status and Trends
products.}

\item{pis}{data.frame; predictory importance data rom \code{\link[=load_pis]{load_pis()}}.}

\item{st_extent}{list; \code{st_extent} list for spatiotemporal filtering.
Required, as results are less meaningful over large spatiotemporal extents.}

\item{by_cover_class}{logical; whether to aggregate FRAGSTATS metrics for the
land cover classes into single values for the land cover classes.}

\item{num_top_preds}{integer; how many predictors to show.}

\item{return_top}{logical; wether to returns a vector of the top predictors,
based on the \code{num_top_preds} param.}

\item{pretty_names}{logical; whether to convert cryptic land cover codes to
readable land cover class names.}

\item{print_plot}{logical; whether to print plot, to allow only the return
the top predictors, if desired.}
}
\value{
Plots barplot and/or returns a vector of top predictors.
}
\description{
For all of the available predictors in a single set of species eBird
Status and Trends products, this function makes a bar plot of those relative
importances, from highest to lowest. Many function parameters allow for
customized plots.
}
\examples{
\dontrun{

# download and load example data
dl_path <- tempdir()
sp_path <- download_data("example_data", path = dl_path)
pis <- load_pis(sp_path)

# define a spatioremporal extent
ne_extent <- list(type = "rectangle",
                  lat.min = 40,
                  lat.max = 47,
                  lon.min = -80,
                  lon.max = -70,
                  t.min = 0.425,
                  t.max = 0.475)

plot_pis(path = sp_path, pis = pis, st_extent = ne_extent)

}
}
