% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebirdst-ppms.R
\name{poisson_dev}
\alias{poisson_dev}
\title{Poisson deviance}
\usage{
poisson_dev(obs, pred)
}
\arguments{
\item{obs}{numeric; observed values.}

\item{pred}{numeric; predicted values.}
}
\value{
A named numeric vector with three elements: model deviance, mean
deviance, and deviance explained.
}
\description{
Poisson deviance
}
\examples{
obs <- c(0, 0, 1, 3, 5, 2)
pred <- c(0.5, 0.1, 2.5, 3.3, 5.2, 2.5)
ebirdst:::poisson_dev(obs, pred)
}
