% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regional-stats.R
\name{regional_stats}
\alias{regional_stats}
\title{Calculate summary statistics over a set of regional polygons}
\usage{
regional_stats(path, regions)
}
\arguments{
\item{path}{character; directory that the Status and Trends data for a given
species was downloaded to. This path is returned by \code{ebirdst_download()}
or \code{get_species_path()}.}

\item{regions}{\link{sf} object;}
}
\value{

}
\description{
Given a set of polygons defining regions, calculate a suite of statistics
for the range and abundance within those regions.
}
\examples{
\donttest{
library(sf)

# download example data
path <- ebirdst_download("example_data", tifs_only = FALSE)
# or get the path if you already have the data downloaded
path <- get_species_path("example_data")

# generate an example "region"
ex_region <- st_point(c(-84.6, 44.3)) \%>\%
  st_sfc(crs = 4326) \%>\%
  st_transform(crs = "+proj=sinu") \%>\%
  st_buffer(dist = 10000) \%>\%
  st_sf(region_name = "Buffered Point", geometry = .)


}
}
