% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stemhelper-ppms.R
\name{sample.grid.cell}
\alias{sample.grid.cell}
\title{Internal sample.grid.cell function}
\usage{
sample.grid.cell(xxx, yyy, xlim = c(NA, NA), ylim = c(NA, NA), nx = 64,
  ny = 64, jitter = F, size = 1, replace = F)
}
\description{
Internal sample.grid.cell function
}
