% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stemhelper-loading.R
\name{st_extent_subset}
\alias{st_extent_subset}
\title{Internal function to handle different kinds of st_extent subsetting}
\usage{
st_extent_subset(data, st_extent, use_time = TRUE)
}
\description{
Internal function to handle different kinds of st_extent subsetting
}
