% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stemhelper-loading.R
\name{stack_stem}
\alias{stack_stem}
\title{Load, extend, and stack all weeks of STEM rasters for a given result type}
\usage{
stack_stem(path, variable, year = 2016, res = "high", st_extent = NA,
  add_zeroes = TRUE, use_analysis_extent = TRUE)
}
\arguments{
\item{path}{character; Full path to directory containing the STEM results
for a single species.}

\item{variable}{character; One of: 'abundance_ensemble_support',
'abundance_lower', 'abundance_upper', 'abundance_umean',
and 'occurrence_umean'.}

\item{year}{numeric; Default is 2016. Other years are available at
res = "low" for inter-year variation analysis.}

\item{res}{character; Default is "high". Use "low" for inter-year variation
analysis to load other years of data.}

\item{st_extent}{list; Optional, use to limit the spatial Extent that the
rasters are loaded into. Must set use_analysis_extent to FALSE.}

\item{add_zeroes}{logical; Default is TRUE. Adds predicted and assumed zero
values to the resulting layers. Set to FALSE to turn this behavior off.}

\item{use_analysis_extent}{logical; Default is TRUE. If STEM results were
run for a custom non-global extent, that extent object is stored in the
configuration file. If TRUE, uses that analysis extent for loading the
rasters. If FALSE, rasters are loaded to the full extent of the
\code{template_raster} object.}
}
\value{
RasterStack containing all available weeks of result.
}
\description{
For one of four result variables, loads all available weeks of rasters
(or temporal subset), extends them to the extent of the study
(or to a custom extent), and stacks them into a RasterStack, with zeroes
added (option to turn this off).
}
\examples{
\dontrun{

sp_path <- "path to species STEM results"

raster_stack <- stack_stem(path = sp_path, variable = "abundance_umean")
}
}
