% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{template_raster}
\alias{template_raster}
\title{Template raster for extending the extents of STEM raster layers}
\format{A RasterLayer with pixels having value of 0 that represents the full
spatial extent, the resolution, and the coordinate system of the STEM study
extent.}
\usage{
template_raster
}
\description{
A raster layer that is used to unify the extents of STEM results across weeks
of the year. Currently, each individual STEM result layer (e.g.,
"abundance_umean") is stored by having all NA values trimmed, which results
in raster layers that have different extents for each species for each week
of the year. This template raster provides the full spatial extent of the
study area.
}
\examples{
str(stemhelper::template_raster)

r <- raster()
extended_raster <- extend(r, extent(stemhelper::template_raster))
}
\keyword{datasets}
