% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stemhelper-plotting.R
\name{train_check}
\alias{train_check}
\title{Checks whether loess predictions exist within available training data}
\usage{
train_check(y, train_data, empty_val)
}
\arguments{
\item{y}{vector; Row of input data. Mostly used in apply() situation.}

\item{train_data}{data.frame; Training data used for loess.}

\item{empty_value}{A replacement value to use to replace loess predictions.}
}
\value{
Either the loess predictions or the provided \code{empty_value}.
}
\description{
Internal function used by \code{loess_fit_and_predict()} in \code{cake_plot()} that
checks to see if there is available training data to support loess
predictions. This is especially valuable for migration gaps when a species
is absence for a few weeks, but loess fills in.
}
\examples{
\dontrun{
# from loess_fit_and_predict()
# do training data check and replace for gaps, edges, and corners
results$preds <- apply(results,
                       1,
                       train_check,
                       empty_val = empty_val,
                       train_data = D)
}
}
\keyword{interal}
