% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trends.R
\name{trends_palette}
\alias{trends_palette}
\title{eBird Trends color palette for mapping}
\usage{
trends_palette(n)
}
\arguments{
\item{n}{integer; the number of colors to be in the palette.}
}
\value{
A character vector of hex color codes.
}
\description{
A diverging red to blue color palette for mapping eBird Trends. Reds are
typically assigned to declining trends and blues for increasing trends.
}
\examples{
trends_palette(9)
}
