% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zero_es_stack}
\alias{zero_es_stack}
\title{RasterStack of zero values based on checklist-level ensemble support}
\format{A RasterStack, where each layer is a week of the year and where the values range of 0 to 100 representing the number of contributing independent models (stixels) with sufficient checklists to fit and predict within that indepdendent model (stixel). Each RasterLayer in the RasterStack has the full spatial extent and same resolution and coordinate system of the STEM study extent.}
\usage{
zero_es_stack
}
\description{
A RasterStacks (one for each week of the year) indicating the ensemble support based on sufficient checklists to fit models within an individiual stixel. Values range from 0 to 100. For presentation, we use a value of 99 to indicate sufficient information to assume that, if there are no positive occurrence or abundance predictions within a given raster cell, that there is a value of 0 occurrence and 0 abundance.
}
\examples{
zero_es_stack
}
\keyword{datasets}
