% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebirdst-mapping.R
\name{abundance_palette}
\alias{abundance_palette}
\title{eBird Status and Trends color palettes for abundance data}
\usage{
abundance_palette(n, season = c("weekly", "breeding", "nonbreeding",
  "migration", "prebreeding_migration", "postbreeding_migration",
  "year_round"))
}
\arguments{
\item{n}{integer; the number of colors (≥ 1) to be in the palette.}

\item{season}{character; the season to generate colors for or "weekly" to
get the color palette used in the weekly abundance animations.}
}
\value{
A character vector of color hex codes.
}
\description{
Generate the color palettes used in the eBird Status and Trends relative
abundance maps.
}
\examples{
# breeding season color palette
abundance_palette(10, season = "breeding")
}
