% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebirdst-ppms.R
\name{binom_test_p}
\alias{binom_test_p}
\title{Binomial test for ensemble support}
\usage{
binom_test_p(x)
}
\arguments{
\item{x}{numeric; named numeric vector with values for \code{pat}, \code{pi_es}, and
\code{pat_cutoff}.}
}
\value{
A numeric p-value.
}
\description{
Binomial test for ensemble support
}
\examples{
ebirdst:::binom_test_p(c(pat = 0.1, pi_es = 75))
}
