% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebirdst-ppms.R
\name{binom_test_p}
\alias{binom_test_p}
\title{Binomial test for ensemble support}
\usage{
binom_test_p(x, pat_cutoff = 1/10)
}
\arguments{
\item{x}{numeric; named numeric vector with values for \code{"pat"} and \code{"pi_es"}.}

\item{pat_cutoff}{numeric; percent above threshold cutoff}
}
\value{
A numeric p-value.
}
\description{
Binomial test for ensemble support
}
\examples{
ebirdst:::binom_test_p(c(pat = 0.1, pi_es = 75))
}
