% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stemhelper-plotting.R
\name{cake_plot}
\alias{cake_plot}
\title{Make a cake plot of the pis and pds}
\usage{
cake_plot(path, pis, pds, st_extent, by_cover_class = TRUE)
}
\description{
Make a cake plot of the pis and pds
}
