% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebirdst-mapping.R
\name{calc_bins}
\alias{calc_bins}
\title{Calculates bins (breaks) based on standard deviations of log-transformed
data for mapping}
\usage{
calc_bins(x)
}
\arguments{
\item{x}{Raster* object; usually from stem_stack() function.}
}
\value{
A vector containing the break points of bins.
}
\description{
Mapping species abundance across the full-annual cycle presents a challenge,
in that patterns of concentration and dispersion in abundance change
throughout the year, making it difficult to define color bins that suit all
seasons and accurately reflect the detail of abundance predictions. To
address this, we selected a method (described by Maciejewski et al. 2013)
that log transforms the entire year of non-zero data, constructs bins with
the log-transformed data using standard-deviations, and then untransforms
the bins.
}
\examples{
\dontrun{

sp_path <- "path to species STEM results"
raster_stack <- stack(sp_path)
year_bins <- calc_bins(raster_stack)

raster::plot(raster_stack[[1]], xaxt = 'n', yaxt = 'n', breaks = year_bins)
}
}
\references{
Ross Maciejewski, Avin Pattah, Sungahn Ko, Ryan Hafen, William S. Cleveland, David S. Ebert.  Automated Box-Cox Transformations for Improved Visual Encoding. IEEE Transactions on Visualization and Computer Graphics, 19(1): 130-140, 2013.
}
