% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{calc_bins}
\alias{calc_bins}
\title{Calculates bins based on standard deviations of log-transformed data}
\usage{
\code{calc_bins(stack))}
}
\arguments{
\item{stack}{A RasterStack object, of abundance results.}
}
\value{
vector containing break points of bins
}
\description{
Mapping species abundnace across the full-annual cycle presents a challenge, in that patterns of concentration and dispersion in abundance change throughout the year, making it difficult to define color bins that suit all seasons and accurately reflect the detail of abundance predictions. To address this, we selected a method (described by Maciejewski et al. 2013) that log transforms the entire year of data, constructs bins with the log-transformed data using standard-deviations, and then untransforms the bins.
}
\examples{
tif_path <- "~"
raster_stack <- stack_stem(tif_path)
year_bins <- calc_bins(raster_stack)

raster::plot(raster_stack[[1]], xaxt='n', yaxt='n', breaks=year_bins)
}
