% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{calc_effective_extent}
\alias{calc_effective_extent}
\title{Plot extent of estimation calculated from subset of centroids}
\usage{
calc_effective_extent(pis, st_extent)
}
\description{
Plot extent of estimation calculated from subset of centroids
}
