% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebirdst-ppms.R
\name{compute_ppms}
\alias{compute_ppms}
\title{Computes the Predictive Performance Metrics for a spatiotemporal extent}
\usage{
compute_ppms(path, st_extent = NA)
}
\arguments{
\item{path}{character; Full path to single species STEM results.}

\item{st_extent}{list; st_extent list for spatiotemporal filtering.}
}
\value{
list of data.frames: binary_stats, occ_stats, count_stats
}
\description{
Loads test data and ensemble support values and then calculates the
predictive performance metrics within a spatiotemporal extent defined by
\code{st_extent}. Use this function directly to access the computed metrics, or
use \code{plot_all_ppms()} or \code{plot_binary_by_time()} to summarize the metrics.
}
\examples{
\dontrun{

sp_path <- "path to species STEM results"

ne_extent <- list(type = "rectangle",
                  lat.min = 40,
                  lat.max = 47,
                  lon.min = -80,
                  lon.max = -70,
                  t.min = 0.425,
                  t.max = 0.475)

ppms <- compute_ppms(path = sp_path, st_extent = ne_extent)
}
}
