% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebirdst-loading.R
\name{download_data}
\alias{download_data}
\title{Download eBird Status and Trends Data}
\usage{
download_data(species, path = rappdirs::user_data_dir("ebirdst"),
  force = FALSE)
}
\arguments{
\item{species}{character; a single six-letter species code (e.g. rufhum).
The full list of valid codes is can be viewed in the \link{runs_w_names} data
frame included in this package. To download the example dataset, use
"example_data".}

\item{path}{character; directory to download the data to. All downloaded
files will be placed in a sub-directory of this directory named according
to the unique run ID associated with this species. Defaults to a persistent
data directory, which can be found by calling
rappdirs::user_data_dir("ebirdst")).}

\item{force}{logical; if the data have already been downloaded, should a
fresh copy be downloaded anyway.}
}
\value{
Path to the run-specific root of the downloaded files.
}
\description{
Download an eBird Status and Trends data package for a single species, or for
an example species, to a specified path. The example data consist of the
results for Yellow-bellied Sapsucker subset to Michigan and are much smaller
than the full dataset making these data quicker to download and process.
}
\examples{
\dontrun{

download_data("example_data", force = TRUE)

}
}
