% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebirdst-habitat.R
\name{ebirdst_habitat}
\alias{ebirdst_habitat}
\alias{plot.ebirdst_habitat}
\title{eBird Status and Trends predictive habitat associations}
\usage{
ebirdst_habitat(path, ext, n_predictors = 15, pland_only = TRUE)

\method{plot}{ebirdst_habitat}(x, date_range, by_cover_class = TRUE, ...)
}
\arguments{
\item{path}{character; Full path to single species STEM results.}

\item{ext}{\link{ebirdst_extent} object; the spatiotemporal extent over which to
calculate the habitat associations. Note that \strong{temporal component of \code{ext}
is ignored is this function}, habitat associations are always calculated
for the full year.}

\item{n_predictors}{integer; number of land cover classes to estimate habitat
associations for.}

\item{pland_only}{logical; For each land cover class, two FRAGSTATS metrics
are used as covariates: the percent of land cover (PLAND) and the edge
density (ED). By default only PLAND covariates are used to estimate habitat
associates. Use \code{pland_only = FALSE} to use both PLAND and ED metrics.}

\item{x}{\link{ebirdst_habitat} object; habitat relationships as calculated by
\code{\link[=ebirdst_habitat]{ebirdst_habitat()}}.}

\item{date_range}{the range of dates for plotting; a 2-element vector of the
start and end dates of the date range, provided either as dates (Date
objects or strings in ISO format "YYYY-MM-DD") or numbers between 0 and 1
representing the fraction of the year. When providing dates as a string,
the year can be omitted (i.e. "MM-DD"). \strong{Leave the argument blank to plot
the full year of data.}}

\item{by_cover_class}{logical; whether to aggregate the FRAGSTATS metrics
(PLAND and ED) for the land cover classes into single values for the land
cover classes. This will also group the 5 types of roads into a single
class.}

\item{...}{ignored.}
}
\value{
An \code{ebirdst_habitat} object, consisting of a data frame giving the
predictor importance and directionality for each predictor for each week of
the year. The columns are:
\itemize{
\item \code{predictor}: the name of the predictor
\item \code{date}: the week centroid expressed as a continuous value between 0-1.
See \link{ebirdst_weeks} to convert these values to ISO dates.
\item \code{importance}: the relative importance of the predictor, these values are
scaled so they sum to 1 within each week.
\item \code{direction}: the direction of the relationship, either 1 for a positive
relationship, -1 for a negative relationship, or NA when the direction of
the relationship is not significant.
}
}
\description{
Combine the predictor importance (PI) and partial dependence (PD) data to
provide an estimate of the importance and directionality of the land cover
classes (i.e. habitat) used as model covariates. The \code{plot()} method can then
be used to produce a cake plot, a stacked area chart showing habitat
associations in which area indicates the importance of a given land cover
class and the position above or below the x-axis indicates the direction of
the relationship. \strong{Note:} This is one of, if not the most, computationally
expensive operations in the package.
}
\examples{
\donttest{
# download example data
path <- ebirdst_download("example_data", tifs_only = FALSE)
# or get the path if you already have the data downloaded
path <- get_species_path("example_data")

# define a spatial extent to calculate ppms over
bb_vec <- c(xmin = -86, xmax = -83, ymin = 42.5, ymax = 44.5)
e <- ebirdst_extent(bb_vec)

# compute habitat associations
habitat <- ebirdst_habitat(path = path, ext = e)
print(habitat)
# produce a cake plot
plot(habitat)
}
}
