% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_species}
\alias{get_species}
\title{Get eBird species code for a set of species}
\usage{
get_species(x)
}
\arguments{
\item{x}{character; vector of species codes, common names, and/or scientific
names.}
}
\value{
A character vector of eBird species codes.
}
\description{
Give a vector of species codes, common names, and/or scientific names, return
a vector of 6-letter eBird species codes. This function will only look up
codes for species for which eBird Status and Trends results exist.
}
\examples{
get_species(c("Black-capped Chickadee", "Poecile gambeli", "carchi"))
}
