% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebirdst-loading.R
\name{label_raster_stack}
\alias{label_raster_stack}
\title{Labels 52 week RasterStack with the dates for each band}
\usage{
label_raster_stack(raster_data)
}
\arguments{
\item{raster_data}{RasterStack or RasterBrick; original eBird Science Product
raster GeoTiff with 52 bands, one for each week}
}
\value{
A RasterStack or Rasterbrick with names assigned for the dates in
the format of "XYYY.MM.DD" per raster package constraints.
}
\description{
The raster package does not allow layer names to be saved with the bands of a
multi-band GeoTiff. Accordingly, all eBird Science Product raster results
cover the entire 52 week temporal extent of analysis. For convenience, this
function labels the RasterStack once it has been loaded in R with the dates
for each band.
}
\examples{
\dontrun{

rs <- stack("/path to GeoTiff/")
rs

rs_labeled <- label_raster_stack(raster_data)
rs_labeled
}
}
