% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebirdst-loading.R
\name{load_config}
\alias{load_config}
\title{Config file loader}
\usage{
load_config(path)
}
\arguments{
\item{path}{character; Full path to the directory containing single species
eBird Status and Trends products.}
}
\value{
environment object containing all run parameters.
}
\description{
Internal function used by load_summary(), load_pis(), and load_pds() to get
configuration variables from eBird Status and Trends products
(from *_config.RData).
}
\examples{
\dontrun{

sp_path <- "path to species eBird Status and Trends products"

e <- load_config(sp_path)
}
}
\keyword{internal}
