% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stemhelper-loading.R
\name{load_config}
\alias{load_config}
\title{Config file loader}
\usage{
load_config(path)
}
\description{
Internal function used by load_summary(), load_pis(), and load_pds() to get
configuration variables from species run information.
}
