% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebirdst-loading.R
\name{load_config}
\alias{load_config}
\title{Config file loader}
\usage{
load_config(path)
}
\arguments{
\item{path}{character; full path to the directory containing single species
eBird Status and Trends products.}
}
\value{
environment object containing all run parameters.
}
\description{
Internal function used by load_summary(), load_pis(), and load_pds() to get
configuration variables from eBird Status and Trends products
(from *_config.RData).
}
\examples{
\dontrun{

# download example data
dl_path <- tempdir()
sp_path <- download_data("example_data", path = dl_path)

# load configuration file
e <- ebirdst:::load_config(sp_path)
e

}
}
\keyword{internal}
