% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebirdst-loading.R
\name{load_config}
\alias{load_config}
\title{Load configuration file}
\usage{
load_config(path)
}
\arguments{
\item{path}{character; directory that the Status and Trends data for a given
species was downloaded to. This path is returned by \code{ebirdst_download()}
or \code{get_species_path()}.}
}
\value{
environment object containing all run parameters.
}
\description{
Internal function used by load_summary(), load_pis(), and load_pds() to get
configuration variables from STEM species run information
(from *_config.RData).
}
\examples{
\donttest{
# download example data
sp_path <- ebirdst_download("example_data", tifs_only = FALSE)
ebirdst:::load_config(sp_path)
}
}
\keyword{internal}
