% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebirdst-loading.R
\name{load_pds}
\alias{load_pds}
\title{Load partial dependency data for the Status and Trends predictors}
\usage{
load_pds(path, model = c("occurrence", "abundance"), return_sf = FALSE)
}
\arguments{
\item{path}{character; directory that the Status and Trends data for a given
species was downloaded to. This path is returned by \code{ebirdst_download()}
or \code{get_species_path()}.}

\item{return_sf}{logical; whether to return an \link{sf} object of spatial points
rather then the default data frame.}
}
\value{
data.frame containing predictor importance values for each stixel, as
well as stixel summary information.
}
\description{
Partial dependence (PD) plots depict the relationship between the modeled
response (occurrence or relative abundance) and the predictor variables used
in the model. Status and Trends provides the data to generate these plots for
every stixel.
}
\examples{
# download example data
sp_path <- ebirdst_download("example_data", tifs_only = FALSE)

# load predictor dependencies
pds <- load_pds(sp_path)

# plot the top 15 predictor importances
# define a spatiotemporal extent to plot data from
bb_vec <- c(xmin = -86.6, xmax = -82.2, ymin = 41.5, ymax = 43.5)
e <- ebirdst_extent(bb_vec, t = c("05-01", "05-31"))
plot_pds(pds, ext = e, n_top_pred = 15, by_cover_class = TRUE)
}
