% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebirdst-loading.R
\name{load_pds}
\alias{load_pds}
\title{Load partial dependencies for single species STEM results}
\usage{
load_pds(path)
}
\arguments{
\item{path}{character; Full path to the directory containing single species
STEM results.}
}
\value{
data.frame containing partial dependency values for each stixel,
as well as stixel summary information.
}
\description{
Loads the partial dependency data (from pd.txt), joins with stixel summary
data, sets the names from \code{load_config()}, and cleans up the data.frame.
This is one of the slower functions in the package, due to the size of the
pd.txt file (usually multiple GB).
}
\examples{
\dontrun{

sp_path <- "path to species STEM results"

pds <- load_pds(sp_path)
}
}
