% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebirdst-loading.R
\name{load_pis}
\alias{load_pis}
\title{Load predictor importances for single species eBird Status and Trends products}
\usage{
load_pis(path)
}
\arguments{
\item{path}{character; Full path to the directory containing single species
eBird Status and Trends products.}
}
\value{
data.frame containing predictor importance values for each stixel, as
well as stixel summary information.
}
\description{
Loads the predictor importance data (from pi.txt), joins with stixel summary
data, sets the names from \code{\link[=load_config]{load_config()}}, and cleans up the data.frame.
}
\examples{
\dontrun{

# download example data
dl_path <- tempdir()
sp_path <- download_data("example_data", path = dl_path)

# load predictor importance
pis <- load_pis(sp_path)
}
}
