% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebirdst-loading.R
\name{load_pis}
\alias{load_pis}
\title{Load predictor importances for single species STEM results}
\usage{
load_pis(path)
}
\arguments{
\item{path}{character; Full path to the directory containing single species
STEM results.}
}
\value{
data.frame containing predictor importance values for each stixel,
as well as stixel summary information.
}
\description{
Loads the predictor importance data (from pi.txt), joins with stixel summary
data, sets the names from \code{load_config()}, and cleans up the data.frame.
}
\examples{
\dontrun{

sp_path <- "path to species STEM results"

pis <- load_pis(sp_path)
}
}
