% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebirdst-loading.R
\name{load_summary}
\alias{load_summary}
\title{Stixel summary file loader}
\usage{
load_summary(path)
}
\arguments{
\item{path}{character; full path to the directory containing single species
eBird Status and Trends products.}
}
\value{
data.frame containing stixel summary information about each stixel
centroid.
}
\description{
Internal function used by \code{\link[=load_pis]{load_pis()}} and \code{\link[=load_pds]{load_pds()}} to get the stixel
summary information (from summary.txt).
}
\examples{
\dontrun{

# download example data
dl_path <- tempdir()
sp_path <- download_data("example_data", path = dl_path)

# stixel summaries
summaries <- ebirdst:::load_summary(sp_path)
}
}
\keyword{internal}
