% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebirdst-loading.R
\name{load_summary}
\alias{load_summary}
\title{Stixel summary file loader}
\usage{
load_summary(path)
}
\arguments{
\item{path}{character; Full path to the directory containing single species
STEM results.}
}
\value{
data.frame containing stixel summary information about each stixel
centroid.
}
\description{
Internal function used by load_pis() and load_pds() to get the stixel
summary information (from summary.txt).
}
\examples{
\dontrun{

sp_path <- "path to species STEM results"

summaries <- load_summary(sp_path)
}
}
\keyword{internal}
