% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_trends}
\alias{load_trends}
\title{Load eBird Trends Data Products for a set of species}
\usage{
load_trends(run_names = NULL, path = get_trends_path())
}
\arguments{
\item{run_names}{character; one or more trends run names, which are a
combination of species code and season (e.g. "woothr_breeding"). The full
list of valid run names can be viewed in the \link{ebirdst_trends_runs} data
frame included in this package. Set this to NULL to load all species.}

\item{path}{character; path to the eBird trends data parquet file, which can
be found with \code{\link[=get_trends_path]{get_trends_path()}}.}
}
\value{
A data frame containing the trends estimates for a set of species.
The following columns are included:
\itemize{
\item \code{srd_id}: unique integer identifier for the grid cell.
\item \code{longitude/latitude}: longitude and latitude of the grid cell center.
\item \code{abd}: relative abundance estimate for the middle of the trend time
period (e.g. 2014 for a 2007-2021 trend).
\item \code{abd_ppy}: the median estimated percent per year change in relative
abundance.
\item \code{abd_ppy_lower/abd_ppy_upper}: the 80\% confidence interval for the
estimated percent per year change in relative abundance.
\item \code{abd_ppy_nonzero}: a logical (TRUE/FALSE) value indicating if the 80\%
confidence limits overlap zero (FALSE) or don't overlap zero (TRUE)
\item \code{abd_trend}: the median estimated cumulative change in relative
abundance over the trend time period.
\item \code{abd_trend_lower/abd_trend_upper}: the 80\% confidence interval for the
estimated cumulative change in relative abundance over the trend time period.
}
}
\description{
Load the eBird Trends data for a single species or a set of species. Ttrends
are estimated on a 27km by 27km grid for a single season per species
(breeding, non-breeding, or resident). For further details on the methodology
consult Fink et al. 2023.
}
\examples{
\dontrun{
# download the trends data
download_ebird_trends()

# load breeding season trends for sage thrasher and sagebrush sparrow
trends <- load_trends(c("sagthr_breeding", "sagspa1_breeding"))
}
}
\references{
Fink, D., Johnston, A., Strimas-Mackey, M., Auer, T., Hochachka, W. M.,
Ligocki, S., Oldham Jaromczyk, L., Robinson, O., Wood, C., Kelling, S., &
Rodewald, A. D. (2023). A Double machine learning trend model for citizen
science data. Methods in Ecology and Evolution, 00, 1–14.
https://doi.org/10.1111/2041-210X.14186
}
