% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebirdst-ppms.R
\name{plot_all_ppms}
\alias{plot_all_ppms}
\title{Plot all predictive performance metrics}
\usage{
plot_all_ppms(path, st_extent)
}
\arguments{
\item{path}{character; Full path to single species STEM results.}

\item{st_extent}{list; st_extent list for spatiotemporal filtering.}
}
\value{
Plot of metric box plots by category
}
\description{
For a spatiotemporal extent, plots bar plots for all available predictive
performance metrics within three sets: Binary Occurrence, Occurrence
Probability, and Abundance.
}
\examples{
\dontrun{

sp_path <- "path to species STEM results"

ne_extent <- list(type = "rectangle",
                  lat.min = 40,
                  lat.max = 47,
                  lon.min = -80,
                  lon.max = -70,
                  t.min = 0.425,
                  t.max = 0.475)

plot_all_ppms(path = sp_path, st_extent = ne_extent)
}
}
