% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebirdst-ppms.R
\name{plot_binary_by_time}
\alias{plot_binary_by_time}
\title{Plot binary occurrence metrics by time}
\usage{
plot_binary_by_time(path, metric, n_time_periods = 52, st_extent = NA)
}
\arguments{
\item{path}{character; Full path to single species STEM results.}

\item{metric}{character; One of: "Kappa", "AUC", "Sensitivity",
"Specificity".}

\item{n_time_periods}{int; Greater than 1 (e.g., 52 for weeks).}

\item{st_extent}{list; st_extent list for spatiotemporal filtering.}
}
\value{
Plot of metric box plots by time.
}
\description{
For a specified number of time periods (ideally weeks or months), plots one
of four (Kappa, AUC, Sensitivity, Specificity) box plots. Use default,
without \code{st_extent} to see range-wide predictive performance, or provide
an \code{st_extent} to see performance within spatiotemporal extent.
}
\examples{
\dontrun{

sp_path <- "path to species STEM results"


plot_binary_by_time(path = sp_path, metric = "Kappa", n_time_periods = 12)
}
}
