% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stemhelper-plotting.R
\name{plot_pds}
\alias{plot_pds}
\title{Plot PDs}
\usage{
plot_pds(pd_name, pds, st_extent, pointwise_pi = TRUE, stixel_pds = TRUE,
  k.cont.res = 25, gbm.n.trees = 500, nnn.bs = 100,
  equivalent.ensemble.ss = 10, ci.alpha = 0.05, mean.all.data = FALSE)
}
\description{
Plot PDs
}
