% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stemhelper-plotting.R
\name{plot_pis}
\alias{plot_pis}
\title{Plotting PIs as barplots}
\usage{
plot_pis(path, pis, st_extent, by_cover_class = FALSE, num_top_preds = 50,
  return_top = FALSE, pretty_names = TRUE, print_plot = TRUE)
}
\description{
Plotting PIs as barplots
}
