% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stemhelper-loading.R
\name{raster_st_subset}
\alias{raster_st_subset}
\title{Spatiotemporal subsetter for STEM result data objects}
\usage{
raster_st_subset(raster_data, st_extent)
}
\arguments{
\item{st_extent}{list; st_extent list object}

\item{data}{RasterStack or RasterBrick; one of four GeoTiff files provided
with results}
}
\value{
Subset of input data as same type, with layers labeled with dates
from label_raster_stack().
}
\description{
Internal function that takes a RasterStack or RasterBrick and a st_extent
list and returns a spatiotemporal subset of the data object. Currently
designed to handle either a 'rectangle' as defined by a lat/lon bounding
box or a 'polygon' as defined by a SpatialPolygon* object. The function
will use temporal information if provided. The t.min and t.max objects in
the \code{st_extent} list are currently able to wrap time around the year
(e.g., t.min = 0.9 and t.max = 0.1 is acceptable). This function will also
returned a date labeled RasterStack or RasterBrick, so there is no need
to use the label_raster_stack() directly if subsetting.
}
\examples{
\dontrun{

rs <- stack("/path to GeoTiff")
rs

# define st_extent list
ne_extent <- list(type = "rectangle",
                  lat.min = 40,
                  lat.max = 47,
                  lon.min = -80,
                  lon.max = -70,
                  t.min = 0.5,
                  t.max = 0.6)

# subset stack with extent
rs_sub <- raster_st_subset(rs, st_extent = ne_extent)
}
}
