% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{runs_w_names}
\alias{runs_w_names}
\title{Data frame of species available}
\format{A data frame with 107 rows and 14 variables:
\describe{
\item{SPECIES_CODE}{Six letter eBird code in eBird Taxonomy v2016}
\item{RUN_NAME}{Unique analysis identifier and the top level folder name
for all results}
\item{SCI_NAME}{Scientific name from eBird Taxonomy v2016}
\item{PRIMARY_COM_NAME}{English common name from eBird Taxonomy v2016}
\item{BREEDING_START_DT}{Breeding season start date}
\item{BREEDING_END_DT}{Breeding season start date}
\item{NONBREEDING_START_DT}{Non-breeding season start date}
\item{NONBREEDING_END_DT}{Non-breeding season start date}
\item{POSTBREEDING_MIGRATION_START_DT}{Post-breeding season start date}
\item{POSTBREEDING_MIGRATION_END_DT}{Post-breeding season start date}
\item{PREBREEDING_MIGRATION_START_DT}{Pre-breeding season start date}
\item{PREBREEDING_MIGRATION_END_DT}{Pre-breeding season start date}
\item{YEAR_ROUND_START_DT}{For resident species, the year-round start date}
\item{YEAR_ROUND_END_DT}{For resident species, the year-round end date}
}}
\usage{
runs_w_names
}
\description{
A dataset containing the available species, with SPECIES_CODE, RUN_NAME,
SCI_NAME, and PRIMARY_COM_NAME. In addition, the dates defining the
boundaries of the seasons are provided. These seasons are defined on a
species-specific basis through expert review. For information on the details
of defining seasons, please see the \href{https://ebird.org/science/status-and-trends/faq#seasons}{seasons section of the FAQ}. Note that
missing dates imply that a season failed expert review for that species
within that season.
}
\keyword{datasets}
