% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{stack_stem}
\alias{stack_stem}
\title{Load, extend, and stack all STEM .tif rasters in a directory}
\usage{
\code{stack_stem(path)}
}
\arguments{
\item{path}{Full path to directory containing more than one STEM .tif raster.}
}
\value{
RasterStack object
}
\description{
Takes all of the .tif rasters in a directory at provided path, loads them,
extends them to the extent of the study, and stacks them into a RasterStack.
In practice, this will often be all 52 weeks of a single variable
(e.g., abundance_mean), but the files could be rearranged and stacked as well
(i.e., a single week of a single variable across multiple species).
}
\examples{
tif_path <- "~"
raster_stack <- stack_stem(tif_path)
}
