% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stemhelper-loading.R
\name{stack_stem}
\alias{stack_stem}
\title{Load, extend, and stack all weeks of STEM rasters for a given result type}
\usage{
stack_stem(path, variable, st_extent = NA, add_zeroes = TRUE, weeks = NA,
  use_analysis_extent = TRUE)
}
\arguments{
\item{path}{character; Full path to directory containing the STEM results
for a single species.}

\item{variable}{character; One of: 'abundance_ensemble_support',
'abundance_lower', 'abundance_upper', 'abundance_umean',
and 'occurrence_umean'.}

\item{st_extent}{list; Optional, use to limit the spatial Extent that the
rasters are loaded into. Must set use_analysis_extent to FALSE.}

\item{add_zeroes}{logical; Default is TRUE. Adds predicted and assumed zero
values to the resulting layers. Set to FALSE to turn this behavior off.}

\item{weeks}{numeric; Optional, a single integer or vector of integers from 1
to 52, representing the week(s) of the year to load. Defaults to all
available. Note that if some weeks are not present, they will not be loaded.}

\item{use_analysis_extent}{logical; Default is TRUE. If STEM results were
run for a custom non-global extent, that extent object is stored in the
configuration file. If TRUE, uses that analysis extent for loading the
rasters. If FALSE, rasters are loaded to the full extent of the
\code{template_raster} object.}
}
\value{
RasterStack containing all available weeks of result.
}
\description{
For one of four result types, loads all available weeks of rasters, extends
them to the extent of the study (or to a custom extent), and stacks them
into a RasterStack, with zeroes added (option to turn this off).
}
\examples{
\dontrun{

sp_path <- "path to species STEM results"

raster_stack <- stack_stem(path = sp_path, variable = "abundance_umean")
}
}
